SQL-1044: Refactor, still need to fill out metadata
