+++ core/src/lib.rs
@@ -11,9 +11,11 @@ mod err;
 pub use err::{Error, Result};
 mod fields;
 pub use fields::MongoFields;
+mod col_metadata;
 mod json_schema;
+pub use col_metadata::MongoColMetadata;
 mod query;
-pub use query::{MongoColMetadata, MongoQuery};
+pub use query::MongoQuery;
 mod stmt;
 pub use stmt::MongoStatement;
