+++ odbc/src/api/errors.rs
@@ -1,6 +1,6 @@
 use constants::{
-    INVALID_ATTR_VALUE, NOT_IMPLEMENTED, NO_DSN_OR_DRIVER, OPTION_CHANGED, RIGHT_TRUNCATED,
-    UNABLE_TO_CONNECT, VENDOR_IDENTIFIER,
+    FUNCTION_SEQUENCE_ERROR, INVALID_ATTR_VALUE, NOT_IMPLEMENTED, NO_DSN_OR_DRIVER, OPTION_CHANGED,
+    RIGHT_TRUNCATED, UNABLE_TO_CONNECT, VENDOR_IDENTIFIER,
 };
 use thiserror::Error;
 
@@ -58,6 +58,7 @@ impl ODBCError {
             ODBCError::OptionValueChanged(_, _) => OPTION_CHANGED,
             ODBCError::OutStringTruncated(_) => RIGHT_TRUNCATED,
             ODBCError::MissingDriverOrDSNProperty => NO_DSN_OR_DRIVER,
+            ODBCError::NumColsOnUnexecutedStatement => FUNCTION_SEQUENCE_ERROR,
         }
     }
 
@@ -73,7 +74,8 @@ impl ODBCError {
             | ODBCError::MissingDriverOrDSNProperty
             | ODBCError::OutStringTruncated(_)
             | ODBCError::UnsupportedDriverConnectOption(_)
-            | ODBCError::OptionValueChanged(_, _) => 0,
+            | ODBCError::OptionValueChanged(_, _)
+            | ODBCError::NumColsOnUnexecutedStatement => 0,
             ODBCError::Core(me) => me.code(),
         }
     }
